% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.efs_operations.R
\name{create_tags}
\alias{create_tags}
\title{Creates or overwrites tags associated with a file system}
\usage{
create_tags(FileSystemId, Tags)
}
\arguments{
\item{FileSystemId}{ID of the file system whose tags you want to modify (String). This operation modifies the tags only, not the file system.}

\item{Tags}{Array of \code{Tag} objects to add. Each \code{Tag} object is a key-value pair.}
}
\description{
Creates or overwrites tags associated with a file system. Each tag is a key-value pair. If a tag key specified in the request already exists on the file system, this operation overwrites its value with the value provided in the request. If you add the \code{Name} tag to your file system, Amazon EFS returns it in the response to the DescribeFileSystems operation.
}
\details{
This operation requires permission for the \code{elasticfilesystem:CreateTags} action.
}
\examples{
# This operation creates a new tag for an EFS file system.
create_tags(
  FileSystemId = "fs-01234567",
  Tags = list(
    list(
      Key = "Name",
      Value = "MyFileSystem"
    )
  )
)

}
