% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.efs_operations.R
\name{delete_tags}
\alias{delete_tags}
\title{Deletes the specified tags from a file system}
\usage{
delete_tags(
  FileSystemId = "string",
  TagKeys = list(
    "string"
  )
)
}
\arguments{
\item{FileSystemId}{[required] ID of the file system whose tags you want to delete (String).}

\item{TagKeys}{[required] List of tag keys to delete.}
}
\description{
Deletes the specified tags from a file system. If the \code{DeleteTags} request includes a tag key that does not exist, Amazon EFS ignores it and doesn't cause an error. For more information about tags and related restrictions, see \href{http://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/cost-alloc-tags.html}{Tag Restrictions} in the \emph{AWS Billing and Cost Management User Guide}.
}
\details{
This operation requires permissions for the \code{elasticfilesystem:DeleteTags} action.
}
\examples{
# This operation deletes tags for an EFS file system.
delete_tags(
  FileSystemId = "fs-01234567",
  TagKeys = list(
    "Name"
  )
)

}
