% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.efs_package.R
\docType{package}
\name{paws.efs-package}
\alias{paws.efs}
\alias{paws.efs-package}
\title{paws.efs: Amazon Elastic File System}
\description{
Amazon Elastic File System (Amazon EFS) provides
    simple, scalable file storage for use with Amazon EC2 instances in the
    AWS Cloud. With Amazon EFS, storage capacity is elastic, growing and
    shrinking automatically as you add and remove files, so your
    applications have the storage they need, when they need it. For more
    information, see the [User
    Guide](http://docs.aws.amazon.com/efs/latest/ug/api-reference.html).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
